+++ subsys/mgmt/mcumgr/lib/cmd/fs_mgmt/CMakeLists.txt
@@ -8,7 +8,6 @@
 target_include_directories(MCUMGR INTERFACE include)
 
 zephyr_library_sources(
-	src/zephyr_fs_mgmt.c
 	src/fs_mgmt.c
 )
