+++ subsys/mgmt/mcumgr/lib/cmd/fs_mgmt/src/fs_mgmt.c
@@ -7,20 +7,21 @@
 
 #include <assert.h>
 #include <zephyr/kernel.h>
+#include <zephyr/sys/util.h>
+#include <zephyr/fs/fs.h>
 #include <limits.h>
 #include <string.h>
-#include <zephyr/sys/util.h>
+#include <errno.h>
 #include <zcbor_common.h>
 #include <zcbor_decode.h>
 #include <zcbor_encode.h>
-#include "zcbor_bulk/zcbor_bulk_priv.h"
 #include <zephyr/fs/fs.h>
-#include "mgmt/mgmt.h"
 #include <smp/smp.h>
+#include "mgmt/mgmt.h"
 #include "fs_mgmt/fs_mgmt.h"
-#include "fs_mgmt/fs_mgmt_impl.h"
 #include "fs_mgmt/fs_mgmt_config.h"
 #include "fs_mgmt/hash_checksum_mgmt.h"
+#include "zcbor_bulk/zcbor_bulk_priv.h"
 
 #if defined(CONFIG_FS_MGMT_CHECKSUM_IEEE_CRC32)
 #include "fs_mgmt/hash_checksum_crc32.h"
@@ -29,7 +30,6 @@
 #if defined(CONFIG_FS_MGMT_HASH_SHA256)
 #include "fs_mgmt/hash_checksum_sha256.h"
 #endif
-
 
 #ifdef CONFIG_FS_MGMT_CHECKSUM_HASH
 /* Define default hash/checksum */
@@ -72,11 +72,34 @@ static const struct mgmt_handler fs_mgmt_handlers[];
 static fs_mgmt_on_evt_cb fs_evt_cb;
 #endif
 
+static int fs_mgmt_filelen(const char *path, size_t *out_len)
+{
+	struct fs_dirent dirent;
+	int rc;
+
+	rc = fs_stat(path, &dirent);
+
+	if (rc == -EINVAL) {
+		return MGMT_ERR_EINVAL;
+	} else if (rc == -ENOENT) {
+		return MGMT_ERR_ENOENT;
+	} else if (rc != 0) {
+		return MGMT_ERR_EUNKNOWN;
+	}
+
+	if (dirent.type != FS_DIR_ENTRY_FILE) {
+		return MGMT_ERR_EUNKNOWN;
+	}
+
+	*out_len = dirent.size;
+
+	return 0;
+}
+
 /**
  * Encodes a file upload response.
  */
-static bool
-fs_mgmt_file_rsp(zcbor_state_t *zse, int rc, uint64_t off)
+static bool fs_mgmt_file_rsp(zcbor_state_t *zse, int rc, uint64_t off)
 {
 	bool ok;
 
@@ -88,16 +111,50 @@ fs_mgmt_file_rsp(zcbor_state_t *zse, int rc, uint64_t off)
 	return ok;
 }
 
+static int fs_mgmt_read(const char *path, size_t offset, size_t len,
+			void *out_data, size_t *out_len)
+{
+	struct fs_file_t file;
+	ssize_t bytes_read;
+	int rc;
+
+	fs_file_t_init(&file);
+	rc = fs_open(&file, path, FS_O_READ);
+	if (rc != 0) {
+		return MGMT_ERR_ENOENT;
+	}
+
+	rc = fs_seek(&file, offset, FS_SEEK_SET);
+	if (rc != 0) {
+		goto done;
+	}
+
+	bytes_read = fs_read(&file, out_data, len);
+	if (bytes_read < 0) {
+		goto done;
+	}
+
+	*out_len = bytes_read;
+
+done:
+	fs_close(&file);
+
+	if (rc < 0) {
+		return MGMT_ERR_EUNKNOWN;
+	}
+
+	return 0;
+}
+
 /**
  * Command handler: fs file (read)
  */
-static int
-fs_mgmt_file_download(struct mgmt_ctxt *ctxt)
+static int fs_mgmt_file_download(struct mgmt_ctxt *ctxt)
 {
 	uint8_t file_data[FS_MGMT_DL_CHUNK_SIZE];
 	char path[CONFIG_FS_MGMT_PATH_SIZE + 1];
 	uint64_t off = ULLONG_MAX;
-	size_t bytes_read;
+	size_t bytes_read = 0;
 	size_t file_len;
 	int rc;
 	zcbor_state_t *zse = ctxt->cnbe->zs;
@@ -136,15 +193,14 @@ fs_mgmt_file_download(struct mgmt_ctxt *ctxt)
 	 * length.
 	 */
 	if (off == 0) {
-		rc = fs_mgmt_impl_filelen(path, &file_len);
+		rc = fs_mgmt_filelen(path, &file_len);
 		if (rc != 0) {
 			return rc;
 		}
 	}
 
 	/* Read the requested chunk from the file. */
-	rc = fs_mgmt_impl_read(path, off, FS_MGMT_DL_CHUNK_SIZE,
-						   file_data, &bytes_read);
+	rc = fs_mgmt_read(path, off, FS_MGMT_DL_CHUNK_SIZE, file_data, &bytes_read);
 	if (rc != 0) {
 		return rc;
 	}
@@ -159,11 +215,73 @@ fs_mgmt_file_download(struct mgmt_ctxt *ctxt)
 	return ok ? MGMT_ERR_EOK : MGMT_ERR_EMSGSIZE;
 }
 
+static int fs_mgmt_write(const char *path, size_t offset,
+			 const void *data, size_t len)
+{
+	struct fs_file_t file;
+	int rc;
+	size_t file_size = 0;
+
+	if (offset == 0) {
+		rc = fs_mgmt_filelen(path, &file_size);
+	}
+
+	fs_file_t_init(&file);
+	rc = fs_open(&file, path, FS_O_CREATE | FS_O_WRITE);
+	if (rc != 0) {
+		return MGMT_ERR_EUNKNOWN;
+	}
+
+	if (offset == 0 && file_size > 0) {
+		/* Offset is 0 and existing file exists with data, attempt to truncate the file
+		 * size to 0
+		 */
+		rc = fs_truncate(&file, 0);
+
+		if (rc == -ENOTSUP) {
+			/* Truncation not supported by filesystem, therefore close file, delete
+			 * it then re-open it
+			 */
+			fs_close(&file);
+
+			rc = fs_unlink(path);
+			if (rc < 0 && rc != -ENOENT) {
+				return rc;
+			}
+
+			rc = fs_open(&file, path, FS_O_CREATE | FS_O_WRITE);
+		}
+
+		if (rc < 0) {
+			/* Failed to truncate file */
+			return MGMT_ERR_EUNKNOWN;
+		}
+	} else if (offset > 0) {
+		rc = fs_seek(&file, offset, FS_SEEK_SET);
+		if (rc != 0) {
+			goto done;
+		}
+	}
+
+	rc = fs_write(&file, data, len);
+	if (rc < 0) {
+		goto done;
+	}
+
+done:
+	fs_close(&file);
+
+	if (rc < 0) {
+		return MGMT_ERR_EUNKNOWN;
+	}
+
+	return 0;
+}
+
 /**
  * Command handler: fs file (write)
  */
-static int
-fs_mgmt_file_upload(struct mgmt_ctxt *ctxt)
+static int fs_mgmt_file_upload(struct mgmt_ctxt *ctxt)
 {
 	char file_name[CONFIG_FS_MGMT_PATH_SIZE + 1];
 	unsigned long long len = ULLONG_MAX;
@@ -234,7 +352,7 @@ fs_mgmt_file_upload(struct mgmt_ctxt *ctxt)
 
 	if (file_data.len > 0) {
 		/* Write the data chunk to the file. */
-		rc = fs_mgmt_impl_write(file_name, off, file_data.value, file_data.len);
+		rc = fs_mgmt_write(file_name, off, file_data.value, file_data.len);
 		if (rc != 0) {
 			return rc;
 		}
@@ -255,8 +373,7 @@ fs_mgmt_file_upload(struct mgmt_ctxt *ctxt)
 /**
  * Command handler: fs stat (read)
  */
-static int
-fs_mgmt_file_status(struct mgmt_ctxt *ctxt)
+static int fs_mgmt_file_status(struct mgmt_ctxt *ctxt)
 {
 	char path[CONFIG_FS_MGMT_PATH_SIZE + 1];
 	size_t file_len;
@@ -283,7 +400,7 @@ fs_mgmt_file_status(struct mgmt_ctxt *ctxt)
 	path[name.len] = '\0';
 
 	/* Retrieve file size */
-	rc = fs_mgmt_impl_filelen(path, &file_len);
+	rc = fs_mgmt_filelen(path, &file_len);
 
 	/* Encode the response. */
 	if (rc == 0) {
@@ -308,8 +425,7 @@ fs_mgmt_file_status(struct mgmt_ctxt *ctxt)
 /**
  * Command handler: fs hash/checksum (read)
  */
-static int
-fs_mgmt_file_hash_checksum(struct mgmt_ctxt *ctxt)
+static int fs_mgmt_file_hash_checksum(struct mgmt_ctxt *ctxt)
 {
 	char path[CONFIG_FS_MGMT_PATH_SIZE + 1];
 	char type_arr[HASH_CHECKSUM_TYPE_SIZE + 1] = FS_MGMT_CHECKSUM_HASH_DEFAULT;
@@ -359,7 +475,7 @@ fs_mgmt_file_hash_checksum(struct mgmt_ctxt *ctxt)
 	}
 
 	/* Check provided offset is valid for target file */
-	rc = fs_mgmt_impl_filelen(path, &file_len);
+	rc = fs_mgmt_filelen(path, &file_len);
 
 	if (rc != 0) {
 		return MGMT_ERR_ENOENT;
@@ -478,8 +594,7 @@ static struct mgmt_group fs_mgmt_group = {
 	.mg_group_id = MGMT_GROUP_ID_FS,
 };
 
-void
-fs_mgmt_register_group(void)
+void fs_mgmt_register_group(void)
 {
 	mgmt_register_group(&fs_mgmt_group);
